+++ test/image/image_types/test_insert_images.py
@@ -3,6 +3,7 @@ import sys
 from pathlib import Path
 
 from PIL import Image, ImageDraw, ImageFont
+import pytest
 
 import fpdf
 from test.conftest import assert_pdf_equal
@@ -24,6 +25,37 @@ def test_insert_jpg(tmp_path):
         assert_pdf_equal(pdf, HERE / "image_types_insert_jpg.pdf", tmp_path)
 
 
+@pytest.mark.skipif(
+    sys.platform in ("cygwin", "win32"),
+    reason="Required system libraries to generate JPEG2000 images are a PITA to install under Windows",
+)
+def test_insert_jpg_jpxdecode(tmp_path):
+    pdf = fpdf.FPDF()
+    pdf.compress = False
+    pdf.set_image_filter("JPXDecode")
+    pdf.add_page()
+    file_path = HERE / "insert_images_insert_jpg.jpg"
+    pdf.image(file_path, x=15, y=15, h=140)
+    assert_pdf_equal(pdf, HERE / "image_types_insert_jpg_jpxdecode.pdf", tmp_path)
+
+
+def test_insert_jpg_flatedecode(tmp_path):
+    pdf = fpdf.FPDF()
+    pdf.compress = False
+    pdf.set_image_filter("FlateDecode")
+    pdf.add_page()
+    file_path = HERE / "insert_images_insert_jpg.jpg"
+    pdf.image(file_path, x=15, y=15, h=140)
+    if sys.platform in ("cygwin", "win32"):
+        # Pillow uses libjpeg-turbo on Windows and libjpeg elsewhere,
+        # leading to a slightly different image being parsed and included in the PDF:
+        assert_pdf_equal(
+            pdf, HERE / "image_types_insert_jpg_flatedecode_windows.pdf", tmp_path
+        )
+    else:
+        assert_pdf_equal(pdf, HERE / "image_types_insert_jpg_flatedecode.pdf", tmp_path)
+
+
 def test_insert_png(tmp_path):
     pdf = fpdf.FPDF()
     pdf.compress = False
@@ -31,6 +63,36 @@ def test_insert_png(tmp_path):
     file_path = HERE / "insert_images_insert_png.png"
     pdf.image(file_path, x=15, y=15, h=140)
     assert_pdf_equal(pdf, HERE / "image_types_insert_png.pdf", tmp_path)
+
+
+def test_insert_png_alpha(tmp_path):
+    pdf = fpdf.FPDF()
+    pdf.compress = False
+    pdf.add_page()
+    pdf.set_font("Helvetica", size=30)
+    pdf.cell(w=pdf.epw, h=30, txt="BEHIND")
+    file_path = HERE / "../png_images/ba2b2b6e72ca0e4683bb640e2d5572f8.png"
+    pdf.image(file_path, x=25, y=0, h=40)
+    assert_pdf_equal(pdf, HERE / "image_types_insert_png_alpha.pdf", tmp_path)
+
+
+def test_insert_png_alpha_dctdecode(tmp_path):
+    pdf = fpdf.FPDF()
+    pdf.compress = False
+    pdf.set_image_filter("DCTDecode")
+    pdf.add_page()
+    file_path = HERE / "../png_images/ba2b2b6e72ca0e4683bb640e2d5572f8.png"
+    pdf.image(file_path, x=15, y=15, h=140)
+    if sys.platform in ("cygwin", "win32"):
+        # Pillow uses libjpeg-turbo on Windows and libjpeg elsewhere,
+        # leading to a slightly different image being included in the PDF:
+        assert_pdf_equal(
+            pdf, HERE / "image_types_insert_png_alpha_dctdecode_windows.pdf", tmp_path
+        )
+    else:
+        assert_pdf_equal(
+            pdf, HERE / "image_types_insert_png_alpha_dctdecode.pdf", tmp_path
+        )
 
 
 def test_insert_bmp(tmp_path):